+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed update token for topic reader
 * Marked sessions which creates from `database/sql` driver as supported server-side session balancing
 
 ## v3.37.7